# PERF: MultiIndex.equals for equal length indexes