- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.3.0.rst` file if fixing a bug or adding a new feature.

Avoids densifying the levels of the multiindex to check for equality.

```
from asv_bench.benchmarks.multiindex_object import Equals

b = Equals()
b.setup()
%timeit b.time_equals_deepcopy()

# 56.4 ms ± 1.03 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)   -> main
# 3.77 ms ± 58.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  -> PR
```